Add new Callback to Swiper
Add new Callback to Swiper:
onSwipeStart : function  
onSwipeEnd : funciton

onSwipeStart is called when a gesture has started. Usefull when you want to do something when swipe begin no mater the direction of the swipe.
onSwipeEnd is called at the end of swipping when user has released all touches. Usefull when you want to do something when swipe end no mater the direction of the swipe.